+++ cf/functions.py
@@ -627,9 +627,9 @@ class log_level(ConstantAccess, cfdm.log_level):
 
 
 class regrid_logging(ConstantAccess):
-    """Whether or not to enable ESMPy regridding logging.
+    """Whether or not to enable `ESMF` regridding logging.
 
-    If it is logging is performed after every call to ESMPy.
+    If it is logging is performed after every call to `ESMF`.
 
     :Parameters:
 
@@ -3316,8 +3316,10 @@ def environment(display=True, paths=True):
         "cf": (__version__, _os_path_abspath(__file__)),
     }
     string = "{0}: {1!s}"
-    if paths:  # include path information, else exclude, when unpacking tuple
+    if paths:
+        # Include path information, else exclude, when unpacking tuple
         string += " {2!s}"
+
     out = [
         string.format(dep, *info)
         for dep, info in dependency_version_paths_mapping.items()